refactor(databse): mobe connect_db to ./utils
