+++ src/core/controllers/depends/users/get_me.py
@@ -7,7 +7,7 @@ from fastapi import Depends
 from src.core.controllers.depends.auth.check_token import (
     get_user_id_by_token_access,
 )
-from src.core.controllers.depends.connect_db import get_crud, get_session
+from src.core.controllers.depends.utils.connect_db import get_crud, get_session
 from src.core.controllers.depends.utils.serialize_user import serialize_user
 from src.core.validators import UserProfile
